♻️ :: 도메인 구조 변형
