+++ src/main/java/com/dsm/nms/domain/comment/CommentRepository.java
@@ -1,0 +1,6 @@
+package com.dsm.nms.domain.comment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+}
