+++ src/main/java/com/dsm/nms/domain/image/ImageRepository.java
@@ -1,0 +1,6 @@
+package com.dsm.nms.domain.image;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ImageRepository extends JpaRepository<Image, Integer> {
+}
