+++ src/main/java/com/dsm/nms/domain/notice/NoticeRepository.java
@@ -1,0 +1,6 @@
+package com.dsm.nms.domain.notice;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+}
