+++ src/main/java/com/dsm/nms/domain/notice/noticetarget/NoticeTargetRepository.java
@@ -1,0 +1,6 @@
+package com.dsm.nms.domain.notice.noticetarget;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NoticeTargetRepository extends JpaRepository<NoticeTarget, NoticeTargetId> {
+}
