+++ src/main/java/com/dsm/nms/domain/refreshtoken/RefreshTokenRepository.java
@@ -1,0 +1,8 @@
+package com.dsm.nms.domain.refreshtoken;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
+}
