+++ src/main/java/com/dsm/nms/domain/reply/Reply.java
@@ -1,0 +1,30 @@
+package com.dsm.nms.domain.reply;
+
+import com.dsm.nms.domain.BaseTimeEntity;
+import com.dsm.nms.domain.comment.Comment;
+import com.dsm.nms.domain.teacher.Teacher;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import javax.persistence.*;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Reply extends BaseTimeEntity {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @Column(nullable = false)
+    private String reply;
+
+    @JoinColumn(name = "comment_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Comment comment;
+
+    @JoinColumn(name = "teacher_id")
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Teacher teacher;
+
+}
