+++ src/main/java/com/dsm/nms/domain/reply/ReplyRepository.java
@@ -1,0 +1,6 @@
+package com.dsm.nms.domain.reply;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+}
