+++ src/main/java/com/dsm/nms/domain/student/Grade.java
@@ -1,0 +1,5 @@
+package com.dsm.nms.domain.student;
+
+public enum Grade {
+    FIRST, SECOND, THIRD
+}
