+++ src/main/java/com/dsm/nms/domain/teacher/Department.java
@@ -1,0 +1,15 @@
+package com.dsm.nms.domain.teacher;
+
+public enum Department {
+    교무부,
+    마이스터부,
+    교육연구부,
+    학생생활안전부,
+    사감부,
+    방과후학교부,
+    환경복지부,
+    교육정보부,
+    진로진학상담부,
+    산학협력부,
+    발명교육센터
+}
