--- src/main/java/com/dsm/nms/entity/image/ImageRepository.java
@@ -1,6 +1,0 @@
-package com.dsm.nms.entity.image;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ImageRepository extends JpaRepository<Image, Integer> {
-}
