--- src/main/java/com/dsm/nms/entity/notice/NoticeRepository.java
@@ -1,6 +1,0 @@
-package com.dsm.nms.entity.notice;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
-}
