--- src/main/java/com/dsm/nms/entity/notice/noticetarget/NoticeTargetRepository.java
@@ -1,6 +1,0 @@
-package com.dsm.nms.entity.notice.noticetarget;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NoticeTargetRepository extends JpaRepository<NoticeTarget, NoticeTargetId> {
-}
