--- src/main/java/com/dsm/nms/entity/refreshtoken/RefreshTokenRepository.java
@@ -1,8 +1,0 @@
-package com.dsm.nms.entity.refreshtoken;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
-}
