--- src/main/java/com/dsm/nms/entity/reply/ReplyRepository.java
@@ -1,6 +1,0 @@
-package com.dsm.nms.entity.reply;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
-}
