--- src/main/java/com/dsm/nms/entity/star/StarRepository.java
@@ -1,6 +1,0 @@
-package com.dsm.nms.entity.star;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StarRepository extends JpaRepository<Star, Integer> {
-}
