--- src/main/java/com/dsm/nms/entity/student/Grade.java
@@ -1,5 +1,0 @@
-package com.dsm.nms.entity.student;
-
-public enum Grade {
-    FIRST, SECOND, THIRD
-}
